refactor: Move Notification delivery logic into protocol
This is a non-behavior changing refactor that will make it easier for
future changes to implement different delievery mechanisms for tap
reminders.

At the time of writing, an NSAlert backed mechanism is planned to the
existing Notifications Center backed mechanism.

In addition to allowing more flexibilty, this abstraction is also
helpful from a code organization and readability perspective.