fix: Remove existing notifications before presenting another
The agent's current behavior is to display multiple notifications at
once if the `present()` method is called sequentially without an
intermediate `dismiss()` call.

In theory this can happen if users GPG sign in different terminal tabs
without acting on previous sign requests. In this case, only the latest
notification would be cleared, which is a bug.

This change was prompted by a future commit refactoring `present()` to
be an async method. The `checkedContinuation` API forced better handling
of this scenario since a continuation discarded without resuming log a
warning.