fix: Handle UNNotificationDismissActionIdentifier
This allows the agent process to properly exit in
--gpg-tap-notifier-test-notification mode when the top left (X) button
is clicked for a notification.

Previously this event wasn't handled and the agent lingered around. The
user had to click on the body of the notification for the agent process
to respond and close.

This change should not affect normal usage of the agent.