Close agent in --gpg-tap-notifier-test-notification mode after 3 seconds
If the user ignores test notifications and never acts on them, the agent
can linger around forever. If the user closes the GUI and reopens it,
this presents an opportunity to test another notification, which allows
infinite agent processes to be spawned.

Limiting agents started with `--gpg-tap-notifier-test-notification` to
a 3 second lifetime to prevent this.

Note that this also causes the initial notification permission request
to end prematurely as well. The next commit in this PR addresses this.